Renamed blobHelpers.ts -> blobs.ts (Util), moved Util encoding to Trie
