+++ packages/trie/src/util/encoding.ts
